Isolate registerUser controller tests
